% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchForTaxon.R
\name{ExtractTaxonNames}
\alias{ExtractTaxonNames}
\title{Extract taxon names}
\usage{
ExtractTaxonNames(data)
}
\arguments{
\item{data}{Data.frame from GetAllDataFromMultipleYears}
}
\value{
list of lists of names
}
\description{
Extract taxon names
}
