% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterForString.R
\name{FilterForString}
\alias{FilterForString}
\title{Search for a single string in award titles and abstract}
\usage{
FilterForString(string, data)
}
\arguments{
\item{string}{Term to search for}

\item{data}{Data.frame from GetAllDataFromMultipleYears}
}
\value{
A data.frame with matches.
}
\description{
Search for a single string in award titles and abstract
}
