% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFromFile.R
\name{LoadPackageData}
\alias{LoadPackageData}
\title{Load data saved in inst/extdata}
\usage{
LoadPackageData(min.year = 1959, max.year = 2017)
}
\arguments{
\item{min.year}{First year of data to pull in}

\item{max.year}{Last year of data to to pull in}
}
\description{
Load data saved in inst/extdata
}
