% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchForTaxon.R
\name{SearchForPhylo}
\alias{SearchForPhylo}
\title{Search for grants in the area of systematics}
\usage{
SearchForPhylo(data, terms = c("systematics", "phylogenetics", "taxonomy",
  "tree of life"))
}
\arguments{
\item{data}{Data.frame from GetAllDataFromMultipleYears}

\item{terms}{Vector of terms to look for}
}
\value{
A data.frame with matches.
}
\description{
Search for grants in the area of systematics
}
