% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchForTaxon.R
\name{GetAllGrantDataForClades}
\alias{GetAllGrantDataForClades}
\title{Pull out funding data for a set of clades}
\usage{
GetAllGrantDataForClades(clade.names, clade.other.names = NULL,
  data = LoadPackageData())
}
\arguments{
\item{clade.names}{Vector of clade names}

\item{clade.other.names}{Optional list of vectors with other clade names ("beetle" for Coleoptera for example)}

\item{data}{Data.frame from GetAllDataFromMultipleYears}
}
\value{
List of data.frames with matches.
}
\description{
Pull out funding data for a set of clades
}
