% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchForTaxon.R
\name{GetAllGrantDataForClades}
\alias{GetAllGrantDataForClades}
\title{Pull out funding data for a set of clades}
\usage{
GetAllGrantDataForClades(clade.names, clade.other.names = as.list(rep(NULL,
  length(clade.names))), data = LoadPackageData(), terms = c("systematics",
  "phylogenetics", "taxonomy", "tree of life"))
}
\arguments{
\item{clade.names}{Vector of clade names}

\item{clade.other.names}{List of vectors with other clade names ("beetle" for Coleoptera for example)}

\item{data}{Data.frame from GetAllDataFromMultipleYears}

\item{terms}{Vector of terms to look for}
}
\value{
List of data.frames with matches.
}
\description{
Pull out funding data for a set of clades
}
